# docs/swagger.yaml
basePath: /v2
definitions:
...
  request.CreateProductRequest:
    properties:
      code:
        type: string
      minimal_stock:
        type: integer
      name:
        type: string
      price:
        type: integer
      quantity:
        type: integer
    required:
    - code
    - minimal_stock
    - name
    - price
    - quantity
    type: object
...
  request.SearchParamProductRequest:
    properties:
      name:
        type: string
    type: object
...
  request.UpdateProductRequest:
    properties:
      minimal_stock:
        type: integer
      name:
        type: string
      price:
        type: integer
    required:
    - minimal_stock
    - name
    - price
    type: object
...
  response.ProductFailedResponse:
    properties:
      code:
        type: integer
      code_type:
        type: string
      error_messages:
        items:
          type: string
        type: array
    type: object
  response.ProductResponse:
    properties:
      code:
        type: string
      id:
        type: string
      message:
        type: string
      name:
        type: string
      on_hold_quantity:
        type: integer
      quantity:
        type: integer
      sold_quantity:
        type: integer
    type: object
  response.ProductSearchResponse:
    properties:
      current_page:
        type: integer
      data:
        items:
          $ref: '#/definitions/response.ProductResponse'
        type: array
      limit:
        type: integer
      total_page:
        type: integer
      total_records:
        type: integer
    type: object
  response.ProductSuccessResponse:
    properties:
      code:
        type: integer
      code_type:
        type: string
      error_messages:
        items:
          type: string
        type: array
      result:
        $ref: '#/definitions/response.ProductResponse'
        type: object
    type: object
  response.ProductSuccessWithPagingResponse:
    properties:
      code:
        type: integer
      code_type:
        type: string
      error_messages:
        items:
          type: string
        type: array
      result:
        $ref: '#/definitions/response.ProductSearchResponse'
        type: object
    type: object
host: petstore.swagger.io
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /product:
    post:
      consumes:
      - application/json
      description: Create Product
      operationId: Create Product
      parameters:
      - description: create product info
        in: body
        name: CreateProduct
        required: true
        schema:
          $ref: '#/definitions/request.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ProductFailedResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ProductFailedResponse'
      tags:
      - Product
  /product/{id}:
    get:
      consumes:
      - application/json
      description: Find a Product
      operationId: Find a Product
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ProductFailedResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ProductFailedResponse'
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: Update Product
      operationId: Update Product
      parameters:
      - description: update product info
        in: body
        name: CreateProduct
        required: true
        schema:
          $ref: '#/definitions/request.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ProductFailedResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ProductFailedResponse'
      tags:
      - Product
  /product/find-all:
    post:
      consumes:
      - application/json
      description: Find All Product
      parameters:
      - description: search product info
        in: body
        name: Searchuser
        required: true
        schema:
          $ref: '#/definitions/request.SearchParamProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductSuccessWithPagingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ProductFailedResponse'
      tags:
      - Product
swagger: "2.0"